skp_ai/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
chromadb
openai
trafilatura
aiohttp
rank_bm25
scikit-learn
numpy
pandas
rich
textdistance
prometheus_client
beautifulsoup4
